Started EventManager Update
I would like to mark this as a Milestone to fixing this project and setting it up for future work. Making it easier to complete.
I have implemented the MonoSingleton and the EventManager scripts from CZG to the project. Using them I have also developed a PlayerInputManager (The only issue was making sure it gets set up in the project: NOTE WILL HAVE TO FIX)

Was able to remove the Update function from the Weapon script! Having it rely only on the EventManager to tell it when to fire and release. The next update should be to Remove it from PlayerController, followed by the Customize Ship scripts.