Continued removing Update functions
Firstly, changed all colliders to Box Colliders, and added Rigidbodies to Scrap, Player, Bullets, Enemies and Hazards.
Added old collision code for Bullets and Player into OnTriggerEnter functions so we don't always have to run an update to check these.
Added Listeners to check for when MiscData.Scrap is updated, to update the text for it.
Added Listeners for player input for when player is trying to move their character, instead of updating constantly.